requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
alembic==1.13.2
anyio==4.4.0
authlib==1.2.0
blinker==1.6.2
click==8.1.7
cryptography==41.0.3
docker==5.0.3
docker-compose==1.29.2
email-validator==2.1.0.post1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.5
greenlet==2.0.2
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==2.5.1
kafka-python==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.8
python-dotenv==0.19.0
python-keycloak==4.3.0
PyYAML==5.1
requests==2.25.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==1.26.5
websocket-client==0.32.0
Werkzeug==3.0.3
WTForms==3.1.2


